+++ ol_exception/models/__init__.py
@@ -7,6 +7,8 @@ from . import exception_rule
 from . import mrp_production
 from . import product_state
 from . import product_template
+from . import purchase_line
 from . import res_partner
 from . import sale_order
+from . import stock_move
 from . import stock_picking
